Seaprate note and notes utils.
